Reverted commits to ChunkFinder to close #862
